+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## CHANGELOG FOR 1.1.x
 #### 1.1.0
+- bugfix [#1450](https://github.com/ergonode/backend/issues/1450) Low product collection column query performance for product grid (rprzedzik)
 - refactor [#1438](https://github.com/ergonode/backend/issues/1438) Commands take aggregates objects instead of id object (wfajczyk)
 - feature [#1439](https://github.com/ergonode/backend/issues/1439) Add param file to fixture loader (wfajczyk)
 - feature [#1442](https://github.com/ergonode/backend/issues/1442) Add roles and condition set Columns to transition grid (rprzedzik)